refactor: files and remove unecessary component
