+++ src/App.js
@@ -10,6 +10,7 @@ function App() {
         credentials="nY6NNTZZ6:27b76b9f-18ea-456c-bc5e-3a5263ebc63d"
         url="https://scalr.api.appbase.io"
         dataField={['original_title', 'original_title.search']}
+        renderNoSuggestion={() => <p>No Suggestion</p>}
       />
     </div>
   );
