+++ src/addons/SuggestionItem.jsx
@@ -1,7 +1,6 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
 import Highlight from 'react-highlight-words';
-import Flex from '../styles/Flex';
 
 const highlightStyle = {
   fontWeight: 600,
@@ -11,7 +10,7 @@ const highlightStyle = {
 };
 
 const SuggestionItem = ({ currentValue, suggestion }) => {
-  const { label, value, title, description, image } = suggestion;
+  const { label, value } = suggestion;
   if (label) {
     // label has highest precedence
     return typeof label === 'string' ? (
@@ -25,44 +24,6 @@ const SuggestionItem = ({ currentValue, suggestion }) => {
       </div>
     ) : (
       label
-    );
-  } else if (title || image || description) {
-    return (
-      <Flex alignItems="center" css={{ width: '100%' }}>
-        {image && (
-          <div css={{ margin: 'auto', marginRight: 10 }}>
-            <img
-              src={image}
-              alt=" "
-              height="50px"
-              width="50px"
-              css={{ objectFit: 'contain' }}
-            />
-          </div>
-        )}
-        <Flex
-          direction="column"
-          css={{ width: image ? 'calc(100% - 60px)' : '100%' }}
-        >
-          {title && (
-            <div className="trim">
-              <Highlight
-                searchWords={currentValue.split(' ')}
-                textToHighlight={title}
-                highlightStyle={highlightStyle}
-                css={{ fontSize: '1rem' }}
-              />
-            </div>
-          )}
-          {description && (
-            <Highlight
-              searchWords={currentValue.split(' ')}
-              textToHighlight={description}
-              highlightStyle={highlightStyle}
-            />
-          )}
-        </Flex>
-      </Flex>
     );
   }
   return value;
